requirements.txt
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-api-python-client==2.181.0
python-dotenv==1.0.0
openai==1.106.1
requests==2.32.5